modified controllers to reflect changes in semantic-epa-model
git-svn-id: https://ipesvn.fzi.de/svn-open/icep/semantic-epa-parent@593 aa18e314-994e-4353-9a36-64d395711e41

+++ semantic-epa-esper/src/main/java/de/fzi/cep/sepa/esper/debs/c1/DebsOutput.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class DebsOutput {
 
 	private List<CellData> cellData;
-	private long delay;
+	private double delay;
 	private long pickup_time;
 	private long dropoff_time;
 	
@@ -22,10 +22,10 @@ public class DebsOutput {
 	public void setCellData(List<CellData> cellData) {
 		this.cellData = cellData;
 	}
-	public long getDelay() {
+	public double getDelay() {
 		return delay;
 	}
-	public void setDelay(long delay) {
+	public void setDelay(double delay) {
 		this.delay = delay;
 	}
 	public long getPickup_time() {
